factor our login query method
